fixing various import problems not caught during initial refactor
