+++ website/models/poster.py
@@ -1,6 +1,10 @@
 from django.db import models
 from django.dispatch import receiver
+from django.conf import settings
 from django.db.models.signals import pre_delete, post_save, m2m_changed, post_delete
+
+import datetime
+import os
 
 from .person import Person
 from .project import Project
